requirements.txt
Flask
requests
python-dotenv
azure-cognitiveservices-speech
python-dotenv